requirements.txt
bcrypt==4.0.1
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2
starlette
pydantic
python-decouple
psycopg2-binary